+++ ceammc/CHANGELOG.md
@@ -21,6 +21,8 @@
   - hexbeat method added to ui.radio to set hexbeat pattern when in checklist mode
   - rotate method added to ui.radio: rotation to the left in checklist mode
   - cellsize method added to ui.radio: resize widget by specifying cell size
+  - ssml method added to speech.rhvoice~: support for SSML text markup
+  - read method added to speech.rhvoice~: read text from files
 - new init argument functions:
   - hexbeat(HEX) - expand hex to list of 1 and 0
   - rotate(N ...) - list rotation (N>0: to the left)